+++ src/components/profile.jsx
@@ -1,32 +1,40 @@
-const Profile = () => (
-  <div className="container">
-    <div className="row">
+import { useSelector } from 'react-redux';
 
-      <div className="col-6">
-        <h2>My Mission</h2>
-        <table className="table table-bordered">
-          <tbody>
-            <tr><th>Telstar</th></tr>
-            <tr><th>SES</th></tr>
-            <tr><th>AsiasSat</th></tr>
-            <tr><th>ABS</th></tr>
-          </tbody>
+const Profile = () => {
+  const { rocketStore } = useSelector((state) => state.rocket);
+  return (
+    <div className="container">
+      <div className="row">
 
-        </table>
-      </div>
-      <div className="col-6">
-        <h2>My Rockets</h2>
-        <table className="table table-bordered">
-          <tbody>
-            <tr><th>Falcon 9</th></tr>
-            <tr><th>Falcony Heavy</th></tr>
-            <tr><th>AsiasSat</th></tr>
-            <tr><th>Starship</th></tr>
-          </tbody>
+        <div className="col-6">
+          <h2>My Mission</h2>
+          <table className="table table-bordered">
+            <tbody>
+              <tr><th>Telstar</th></tr>
+              <tr><th>SES</th></tr>
+              <tr><th>AsiasSat</th></tr>
+              <tr><th>ABS</th></tr>
+            </tbody>
 
-        </table>
+          </table>
+        </div>
+        <div className="col-6">
+          <h2>My Rockets</h2>
+          <table className="table table-bordered">
+            <tbody>
+              {rocketStore
+                .filter((rocket) => rocket.reserved)
+                .map((reservedRocket) => (
+                  <tr key={reservedRocket.id}>
+                    <th className="mprtd">{reservedRocket.rocket_name}</th>
+                  </tr>
+                ))}
+            </tbody>
+
+          </table>
+        </div>
       </div>
     </div>
-  </div>
-);
+  );
+};
 export default Profile;
